requirements.txt
streamlit==1.38.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
mahotas==1.4.13
scikit-image==0.22.0
Pillow==10.2.0
lightgbm==4.3.0
scikit-learn==1.5.2
matplotlib==3.8.0
joblib==1.3.2
